Clean-up PNG tests files
